requirements.txt
click
discord.py
requests
python-dotenv
